Update environment variable parsing and action.yml
The `Program.cs` file has been updated to parse the repository owner, repository name, and pull request number from the `GITHUB_REPOSITORY` and `GITHUB_REF_NAME` environment variables, respectively. This change improves the way the program retrieves and processes these details. Additionally, the `action.yml` file has been updated to define the name and description of the GitHub Action, and to specify that it runs using Docker. The Docker image to be used is also specified.