`Removed properties and package from BlogValidator.csproj`
Removed the `<InvariantGlobalization>` and `<PublishTrimmed>` properties from the `BlogValidator.csproj` file. This may affect the application's behavior in different culture settings and increase the size of the published output. Also, the `GitHub.Actions.Octokit` package reference has been removed, which might impact the application's ability to interact with GitHub Actions.